Major Modifications to Support Extended Recording
This set of modifications changes the pipeline to first cluster just the taste delivery interval, to find good taste-related spiking, and then use template matching to find matching spikes throughout the recording. It also changes the way things are run, using 'run_clustering.py' to call functions in order and keep track of the status of sorting a dataset in a csv file in the data folder.